Fix Conclusion

+++ src/controllers/exams/uploadConclusion.ts
@@ -1,40 +1,40 @@
-import { Response, Request } from "express";
-import ErrorUtils, { IErrorData } from "../../utils/constant/Error";
 import csv from "csvtojson";
-import { IConclusion, IConclusionResponse } from "../../utils/interfaces";
+import { Request, Response } from "express";
 import { ConclusionModel } from "../../models/Conclusion";
-import path from "path";
-import { promises as fsPromises } from "fs-extra";
+import ErrorUtils, { IErrorData } from "../../utils/constant/Error";
+import { IConclusion, IConclusionResponse } from "../../utils/interfaces";
 export const uploadConclusion = async (
   req: Request,
   res: Response<IConclusionResponse | IErrorData>
 ) => {
   try {
     const { isReplaceAll = true } = req.query;
-    const conclusions = await csv().fromFile(req.file?.path!);
+
+    // Convert file buffer to string
+    const fileContent = req?.file?.buffer.toString("utf-8");
+
+    // const conclusions = await csv().fromFile(req.file?.path!);
+    // Parse CSV data from the file content
+    if (!fileContent) {
+      return res.send(ErrorUtils.get("SERVER_ERROR"));
+    }
+    const conclusions = await csv().fromString(fileContent);
+
+    //console.log("conclusion", conclusions);
     const validateConclusions = conclusions?.filter(
       (conclusion: IConclusion) => conclusion.Type !== ""
     );
+
     if (JSON.parse(isReplaceAll as string)) {
       await ConclusionModel.deleteMany({});
     }
-    await ConclusionModel.insertMany(validateConclusions).then(async () => {
-      // Delete file in vscode after add in mongodb
-      const directoryPath = path.join(__dirname, "../../../public/uploads");
-      const files = await fsPromises.readdir(directoryPath);
-      for (const file of files) {
-        const filePath = path.join(directoryPath, file);
-        await fsPromises.unlink(filePath);
-        console.log("Successfully deleted file: ", filePath);
-      }
-
-      ///////////////
-      return res.send({
-        code: 200,
-        data: validateConclusions,
-      });
+    await ConclusionModel.insertMany(validateConclusions);
+    return res.send({
+      code: 200,
+      data: validateConclusions,
     });
   } catch (e) {
+    console.log("e", e);
     return res.send(ErrorUtils.get("SERVER_ERROR"));
   }
 };
